Move components files to components directory and delete unnecessary index.js files
