+++ src/components/BoardView.scss
