+++ src/components/CardView.scss
@@ -1,4 +1,4 @@
-.card {
+.game-card {
   perspective: 1500px;
   position: relative;
   height: 150px;
